# Fix stuttering when starting to scroll a Game List