--- a/es-app/src/views/gamelist/VideoGameListView.h
+++ b/es-app/src/views/gamelist/VideoGameListView.h
@@ -33,6 +33,8 @@ private:
 	void initMDLabels();
 	void initMDValues();
 
+	FileData* mFile;
+
 	ImageComponent mThumbnail;
 	ImageComponent mMarquee;
 	VideoComponent* mVideo;