@Gemba 2021-05-22
First: 
Thanks for documenting the variables. +1  Makes sense to me. However, I guess the wording needs some review (e.g. surplus "the"). My suggestion: Why not renaming `mScrollTier` to `mScrollGear` and `mScrollVelocity` to `mScrollStep` (as it is literally a fixed value and not a quotient over time)? This way `mScrollGearAccumulator` would almost explain itself.

Second: 
I found a small regression.

Large gamelist (> 1 Screen):
Single step: Fade out/fade in (differs from HEAD: no fading)
Page step: Fade out/fade in (differs from HEAD: no fading)

Short list (< 1 Screen):
Single step: Fade out/fade in (differs from HEAD: no fading)
Page step: No fading / hard skip (ok as in HEAD)

These are ok:
Continous scrolling / Dimmed with Initials: ok (same as in HEAD)
Roll-over (i.e. pushing up at first list item, pushing down on last list item): ok (same as in HEAD)

The introduced fadeout/fadein for single actions make the ES feel less snappy IMHO.