@XenuIsWatching 2021-05-22
> First:
> Thanks for documenting the variables. +1 Makes sense to me. However, I guess the wording needs some review (e.g. surplus "the"). My suggestion: Why not renaming `mScrollTier` to `mScrollGear` and `mScrollVelocity` to `mScrollStep` (as it is literally a fixed value and not a quotient over time)? This way `mScrollGearAccumulator` would almost explain itself.

`mScrollVelocity` is really the direction you are scrolling. '-1` for `up, `1` for `down`, and `0` for stop scrolling. I could be renaming variables all over the place for clairty :/ ... but I didn't want to go down that rabbit hole. That could be rather a separate pull request, as I want this to stay focused on the issue at hand rather than updating the code for easier maintance and readability.

> Second:
> I found a small regression in the DetailGamelistView.
> 
> Large gamelist (> 1 Screen):
> Single step: Fade out/fade in (differs from master: no fading)
> Page step: Fade out/fade in (differs from master: no fading)

Thanks for testing this out! 
About the single step, please see my comment below.
For the page step, I believe the stutter was so bad, you never saw the fading that was supposed to be happening...

> Short list (< 1 Screen):
> Single step: Fade out/fade in (differs from master: no fading)
> Page step: No fading / hard skip (ok as in master)

Again, Thanks for testing this out! 
Single step, same as above...
Page step, hmm... that's interesting that it was different behavior. I'll try to reproduce it, and investigate why that is happening.

> These are ok:
> Continous scrolling / Dimmed with Initials: ok (same as in master)
> Roll-over (i.e. pushing up at first list item, pushing down on last list item): ok (same as in master)

There is a bug I noticed in master with the Continuous scrolling, and then hitting the very end of the list. When it hits the very end of the list, It will load the info panel, even if the user hasn't let go of the button. But then once the user lets go of the button, it will reload the info panel again. You can notice this if you have a long description, and you hold the button until it begins scrolling after a brief moment. When you let go of it, the description will reset back to beginning indicateing it reloaded.

But this is very minor bug IMO...

> The introduced fadeout/fadein for single actions make the ES feel less snappy IMHO.

I thought this at first, and I did experiment with this using that `isSingleScrolling` function I mentioned earlier. What I did was just force it to retain the state of the info panel when the button was first pressed, and then blank once it is known that the user is holding it down (scrolling by at least 2). But IMHO, I actually felt the opposite and found that this implementation felt less snappier.
My thought went like this....

I press down a button, I see the next item selected... then I see the previous item still on the screen (albeit very briefly) wondering what's going on asking myself "Why am I still seeing old information on the screen, I already see the ?", and then I release the button, I finally see the update info after very brief moment again. All of this happens in less than quarter a second...

For the case I have implemented, my thought went like this...

I press down a button, I see the info panel fade out (and I think to myself it must be doing something and I feel like it's moving), and then I finally release the button, I see the update info after very brief moment again. Again, all of this happens in less than quarter a second...

This is just IMHO as well, but if there is consensus from the maintainers of this project, then I can easier make it perform the former action.