@XenuIsWatching 2021-05-27
I decided to go a simpler route which did no introduce blank in between single scrolling. All this does is record what it just wrote to the info panel, so that when it gets called again (for example, on a button release). it won't waste time updating the panel.

I found that in my system this function can take a while by adding in my own timers ( with `<chrono>`). I compared both my change and the base code. In the base code it calls it twice for button releases. I found that the first call (when the button is first pressed (rising edge), the time for this function can quite a long time under certain conditions in my system (50ms to 1000ms), but the subsequent calls that happened on the button release would take less (5ms to 35ms each). I can only speculate why the subsequent calls are so much shorter, but I would guess there is either some caching going on or my hard drive controller is doing some "magic". These subsequent calls are just rereading the same data and completely redundant and a waste of time.

Logs of my timing are below...
[retropie_infopanel_mod_time.txt](https://github.com/RetroPie/EmulationStation/files/6551491/retropie_infopanel_mod_time.txt)
[retropie_infopanel_base_time.txt](https://github.com/RetroPie/EmulationStation/files/6551492/retropie_infopanel_base_time.txt)