@Gemba 2022-09-02
I tried the suggested change on a installation with the ProfilingUtil (#716). 
As expected there was significant effect measurable (Rpi4 with UASP USB to SATA adapter):

```
es-app/src/views/gamelist/DetailedGameListView.cpp
Commit: 52c04d7
Message          	       Calls	  Total Time	    Avg Time	    Min Time	    Max Time	 Internal Total Time	   Internal Avg Time
updateInfoPanel()	         316	    9.973043	    0.031560	    0.000001	    0.108529	            9.973043	            0.031560

With this PR:
Message          	       Calls	  Total Time	    Avg Time	    Min Time	    Max Time	 Internal Total Time	   Internal Avg Time
updateInfoPanel()	         314	   10.098609	    0.032161	    0.000001	    0.104449	           10.098609	            0.032161
```


Could you pls try on your setup with the NAS? And: Are you using a wired or wireless connection to your NAS?

In a nutshell:
`cmake . -DPROFILING=on -DRPI=on -DGL=on`, `#include "utils/ProfilingUtil.h"` in object under test, then `ProfileScope("...");` in the method to profile. Start ES with `--debug`.